lib/requirements.txt
jinja2
pyyaml
requests
appdaemon